auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv
pydantic[email]
python-multipart